// src/common/enums/ssp-entity-folder.enum.ts
export enum SSPEntityFolderType {
  Application = 'Application',
  Task = 'Task',
  Asset = 'Asset',
  PackageDescriptor = 'PackageDescriptor',
  AvailableActionDescriptor = 'AvailableActionDescriptor',
  AssetDescriptor = 'AssetDescriptor',
  Plugin = 'Plugin',
  Report = 'Report',
  Workflow = 'Workflow',
  Workspace = 'Workspace',
  Dashboard = 'Dashboard'
}